Replace rust with rust2 implementation
