--- rust/src/assets.rs
@@ -1,1 +1,0 @@
-pub const TOOLS_JSON: &str = include_str!("assets/tools.json");
