--- rust/src/sse/mod.rs
@@ -1,3 +1,0 @@
-pub mod claude;
-pub mod openai;
-pub mod toolcall;
